move new function from /job/files/parquet to /service/compact/merge
